fix: restore runtime log worker logger and reduce pkg/factory/service file count
Pass the hosted runtime session logger into LoadWorkerOpts so command-runner
records mirror to the runtime log sink again, and consolidate host, input, and
modelhost files to satisfy the package file-count gate.
